+++ src/main/java/xyz/nikitacartes/easyauth/event/AuthEventHandler.java
@@ -332,8 +332,8 @@ public class AuthEventHandler {
             //? if >= 1.20.3 {
             netHandler.profile = Uuids.getOfflinePlayerProfile(netHandler.profile.getName());
             //?} else if >= 1.20.2 {
-            //netHandler.profile = ServerLoginNetworkHandler.createOfflineProfile(netHandler.profile.getName());
-            //?} else {
+            /*netHandler.profile = ServerLoginNetworkHandler.createOfflineProfile(netHandler.profile.getName());
+            *///?} else {
             /*netHandler.profile = netHandler.toOfflineProfile(netHandler.profile);
             *///?}
         }