+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ##### Minecraft 1.20[.X], 1.21[.X]
 #### Fix
 - Re-register `register` command if `single-use-global-password` was changed
+- Wrong log in message with enabled `single-use-global-password` and disabled `enable-global-password`
 
 ----
 ### 3.3.3